RSX/Remix: first-person viewmodel detection and tagging, plus geometry census
Identifies Resistance 2's first-person arms/weapon and tags them
REMIXAPI_INSTANCE_CATEGORY_BIT_VIEW_MODEL so the Remix runtime routes them
through its viewmodel pass instead of submitting them as ordinary world
geometry - lit by world lights, occluded by world walls, placed by the world
camera. Before this the backend had no notion of a viewmodel at all: the
category bit and REMIXAPI_CAMERA_TYPE_VIEW_MODEL were declared in remix_c.h and
neither appeared anywhere in code, and the Remix dev menu read
"VIEWMODEL Position: -" on every capture.

The rule is the viewport depth range, which is a mechanism rather than a
correlation. The G0 recovery gives z_ndc = 1 - near/w, i.e. z_ndc in [0,1] with
the far plane at infinity, so the viewport maps a draw into window depth
[offset_z, offset_z + scale_z]. R2 draws the whole world at [0, 1] and one thing
at [0, 0.2] - the front fifth - which forces every one of its pixels in front of
any world pixel. That is the standard "the weapon never clips into a wall" trick,
and it is why the bucket exists. Replayed over all 10153 draws in the scratchpad
dumps the rule selects 168, every one [0, 0.2], over exactly two vertex programs.
Those same two programs also draw a 5954-vertex character at [0, 1] in 78 draws,
so the program hash does not separate viewmodel from world geometry and only the
per-draw depth range does - which is why the test is here rather than in
classify_draw() or a hash list. RPCS3_REMIX_VIEWMODEL: 0 off, 1 measure only,
2 measure and tag (default).

Tagging is necessary but not sufficient, and the rest is outside this repo: the
runtime needs a REMIXAPI_CAMERA_TYPE_VIEW_MODEL camera submitted every frame
(not synthesised - the perspective correction is built from both that camera and
the world one), rtx.viewModel.enable = True, and
rtx.playerModel.enableInPrimarySpace = False. The measurement for the camera is
in hand: these draws carry fovx 60.001 / fovy 36.132 with near 0.090 against the
world's 72.000 / 44.634. The comment at the tagging site records this and is now
accurate against the fork, which grew the routing in dxvk-remix-numos3
6476faea (VIEW_MODEL at bit 26, categoryToCameraType -> DrawCallState); it
previously stated the bit was silently dropped, which was true when written.

The viewmodel test is the mirror image of the sky-dome test and is decided in
the same place and for the same reason - the anchor it reports is a statement
about world space, so it cannot be made before per_draw_transform() has resolved
the instance transform. It is therefore interleaved with, and committed
alongside, the rest of that measurement pass:

- RPCS3_REMIX_SKYHASH: identify the sky by albedo content hash rather than by
  geometry, using the dome-shaped predicate only to learn which hashes qualify.
  A hash arms after 8 all-dome draws and is disqualified permanently by one
  non-dome draw. 0 off, 1 measure only (default), 2 tag.
- RPCS3_REMIX_STREAKGATE: refuse a draw whose post-transform bounding box spans
  more than N times the frame's own median world extent. Every audit that
  existed before was scale-free by construction and therefore blind to a mesh
  that is coherent but a thousand times too big.
- RPCS3_REMIX_VTXSPREAD / RPCS3_REMIX_VTXREFUSE: report, and optionally drop,
  draws whose decoded positions do not form one object.
- RPCS3_REMIX_CAT: category tagging census.
- skin_unrec_arl / _foreign / _reads: attribute the skinned-rig refusals that
  had been non-zero and unattributed for this whole title (186648 draws against
  skin_submitted=43698). The reason was already computed in
  vp_fingerprint::skin_note and simply never reported.

Also silences seven C4244 narrowing warnings by casting the cfg::_float reads in
remix_rsx accessors, which return f32 while cfg::_float::float_type is f64.
No config schema change.

+++ rpcs3/Emu/RSX/Remix/RemixRuntime.cpp
@@ -592,7 +592,7 @@ namespace remix_rsx
 			return 0.f;
 		}();
 
-		return env > 1.f ? env : g_cfg.video.remix.far_plane;
+		return env > 1.f ? env : static_cast<f32>(g_cfg.video.remix.far_plane);
 	}
 }
